Rename executable to oscopy_app, FIXED: Put ipythonrc-oscopy in /home/arnaud/.ipython if not existing.
As oscopy name is already existing in the source directory, executable create from ioscopy_app.py cannot be named 'oscopy' so we use the name 'oscopy_app'.

IPython require the file ipythonrc-oscopy to be present in /home/arnaud/.ipython. We test if this file exist to copy it if needed before launching ipython. ipythonrc-oscopy is located in @datarootdir@/oscopy.