FIXED: Exit through GUI menu, rename executable to ioscopy
Per default, ipython -pylab use a multithreaded version of the shell, therefore calling gtk.main_quit() would only exit the gtk thread.
This commit is mainly derived from posts #6 and #7 of https://bugs.launchpad.net/ipython/+bug/270856.

And the name is changed to be shorter and reflect the fact we use oscopy module together with ipython.