Move Figure GUI code in dedicated class IOscopy_GTK_Figure
This class derives from oscopy.Figure and manage the GUI with GTK (the window and its widgets, the contextual menus).

The menuitems from contextual menu use now direct functions from Figure/Graph and not anymore calls to the CLI through "_app_exec" as the IOscopy_GTK_FIgure is not aware of the application.
For the same reason, the "Insert signal" menuitem has been removed. Proposal for new place is contextual menu of the Signal list in the main window.

BUG: Graph deletion is broken, apparently due to the new stack-based Graph management.

Access to last elements of lists have been pythonized (e.g. a_list[-1]) as much as possible