FIXED: Support for drag and drop from main GUI window to Figure
It is not _that_ trivial to keep the selection when user start to drag. The default handler reset the selection when button 1 is pressed. So we use our own handler to store the selection until drag is recognized. The user will see diseappear the selection but it is still there for the DnD.