Append _xxxx to new signals in case of duplicate name
The behavior in case of update is not managed by this patch.
xxxx is a number ranging from 0000 to 9999, eg. vds_0000 for the second occurence of a vds, vds_0001 for the thirds, etc.
To use test/dupl.oscopy, once simulation is performed demo/irf540.dat shall be copied as demo/irf540-2.dat and demo/irf540-3.dat