Prepare signal list for DnD on treeview selection change
This shall allow to drag and drop a signal list on first drop to the gtk_figure.
The selection is reset at mouse release, therefore to DnD multiple signal at once it is needed to NOT release the mouse after the selection of the last signal until the drop in the gtk_figure.